// src/main/global/api/restful/userManagmentAPI/clientsManager/clientsTypes.ts
export type Client = {
  clientId: number;
  firstName: string;
  lastName: string;
  clientType: number;
  clientStatus: number;
  mobile: string | null;
  email: string;
  image: string | null;
  nationality: string | null;
  cityName?: string | null;
  nationalId: string | null;
  copyNum: string | null;
  gender: number;
  cityId: number;
  address: string | null;
  password: string;
  points: number;
  membership: number;
  addedDate: string;
  lastOtp: string | null;
  birthdate: string | null;
  licenseType: string | null;
  licenseExpiration: string | null;
  systemId: string | null;
  notes: string | null;
  membershipName: string | null;
  refCode: string;
  pointsCalculation: string;
  size?: number;
};